Refactor data source URL validation
